Refactor safe_repr and start using it in the safe_signature file
Fix the test to be passing as expected
